Passando arquivos para biblioteca cli-lib
