+++ include/keyboard.h
@@ -1,19 +1,6 @@
-/**
- * keyboard.h
- * Created on Aug, 23th 2023
- * Author: Tiago Barros
- * Based on "From C to C++ course - 2002"
-*/
+#ifndef KEYBOARD_H
+#define KEYBOARD_H
 
-#ifndef __KEYBOARD_H__
-#define __KEYBOARD_H__
+void configurarTerminal(int enable);
 
-void keyboardInit();
-
-void keyboardDestroy();
-    
-int keyhit();
-
-int readch();
-
-#endif /* __KEYBOARD_H__ */
+#endif // KEYBOARD_H
