+++ include/screen.h
@@ -1,181 +1,20 @@
-/**
- * screen.h
- * Created on Aug, 23th 2023
- * Author: Tiago Barros
- * Based on "From C to C++ course - 2002"
- * Reference: https://en.wikipedia.org/wiki/ANSI_escape_code
-*/
+#ifndef SCREEN_H
+#define SCREEN_H
 
-#ifndef __SCREEN_H__
-#define __SCREEN_H__
+#define N 10
+#define M 10
+#define KEY_UP 'w'
+#define KEY_DOWN 's'
+#define KEY_LEFT 'a'
+#define KEY_RIGHT 'd'
 
-#include <stdio.h>
-
-// Terminal control sequences
-#define ESC            "\033"
-#define NORMALTEXT     "[0m"
-#define BOLDTEXT       "[1m"
-#define ITALICTEXT     "[3m"
-#define BLINKTEXT      "[5m"
-#define REVERSETEXT    "[7m"
-#define HOMECURSOR     "[f"
-#define SHOWCURSOR     "[?25h"
-#define HIDECURSOR     "[?25l"
-#define CLEARSCREEN    "[2J"
-
-// BOX Drawing - Unix like terminals
-#define BOX_ENABLE     "(0"
-#define BOX_DISABLE    "(B"
-#define BOX_VLINE      0x78
-#define BOX_HLINE      0x71
-#define BOX_UPLEFT     0x6C
-#define BOX_UPRIGHT    0x6B
-#define BOX_DWNLEFT    0x6D
-#define BOX_DWNRIGHT   0x6A
-#define BOX_CROSS      0x6E
-#define BOX_TLEFT      0X74
-#define BOX_TRIGHT     0X75
-#define BOX_TUP        0X77
-#define BOX_TDOWN      0X76
-
-#define BOX_DIAMOND    0x60
-#define BOX_BLOCK      0x61
-#define BOX_DOT        0x7E
-
-// screen constants
-#define SCRSTARTX      3      // Initial and final screen positions for the game
-#define SCRENDX        75     // It means the area that can be drawn 
-#define SCRSTARTY      1
-#define SCRENDY        23
-
-#define MINX           1      // min screen horizontal pos
-#define MINY           1      // min screen vertical pos
-#define MAXX           80     // max screen horizontal pos
-#define MAXY           24     // max screen vertical pos
-
-/**
- * Screen Colors type
-*/
-typedef enum {BLACK, RED, GREEN, BROWN, BLUE, MAGENTA, CYAN, LIGHTGRAY,
-        DARKGRAY, LIGHTRED, LIGHTGREEN, YELLOW, LIGHTBLUE, 
-        LIGHTMAGENTA, LIGHTCYAN, WHITE} screenColor; 
+int executarFase(int fase);
 
 
+void exibirLabirinto(char labirinto[N][M]);
+void encontrarPosicaoInicial(char labirinto[N][M], int *x, int *y);
+void moverJogador(char labirinto[N][M], int *x, int *y, int tecla);
+char (*obterLabirinto(int fase))[M];
+int executarFase(int fase);
 
-/**
- * Move the cursor to position (0,0)
-*/
-static inline void screenHomeCursor()
-{
-    printf("%s%s", ESC, HOMECURSOR);
-}
-
-/**
- * Show the cursor
-*/
-static inline void screenShowCursor()
-{
-    printf("%s%s", ESC, SHOWCURSOR);
-}
-
-/**
- * Hide the cursor
-*/
-static inline void screenHideCursor()
-{
-    printf("%s%s", ESC, HIDECURSOR);
-}
-
-/**
- * Clear the screen
-*/
-static inline void screenClear()
-{
-    screenHomeCursor();
-    printf("%s%s", ESC, CLEARSCREEN);
-}
-
-/**
- * Update screen imediatelly 
-*/
-static inline void screenUpdate() {
-    fflush(stdout);
-}
-
-/**
- * Set screen mode to "normal"
-*/
-static inline void screenSetNormal()
-{
-    printf("%s%s", ESC, NORMALTEXT);
-}
-
-/**
- * Set screen mode to "bold"
-*/
-static inline void screenSetBold()
-{
-    printf("%s%s", ESC, BOLDTEXT);
-}
-
-/**
- * Set screen mode to "blink"
-*/
-static inline void screenSetBlink()
-{
-    printf("%s%s", ESC, BLINKTEXT);
-}
-
-/**
- * Set screen mode to "reverse"
-*/
-static inline void screenSetReverse()
-{
-    printf("%s%s", ESC, REVERSETEXT);
-}
-
-/**
- * Enable BOX characters in terminal
-*/
-static inline void screenBoxEnable()
-{
-    printf("%s%s", ESC, BOX_ENABLE);
-}
-
-/**
- * Disable BOX characters in terminal
-*/
-static inline void screenBoxDisable()
-{
-    printf("%s%s", ESC, BOX_DISABLE);
-}
-
-/**
- * Clear the screen, set cursor to home position
- * and optionally draw borders on it.
- * 
- * @param drawBorders if not zero, draw borders on screen.
-*/
-void screenInit(int drawBorders);
-
-/**
- * Clear the screen and restores to initial state.
-*/
-void screenDestroy();
-
-/**
- * Move cursor to position (x,y)
- * @param x x position
- * @param y y position
- */ 
-
-void screenGotoxy(int x, int y);
-
-/**
- * Define text colors 
- * @param fg foreground color, can assume values from BLACK to WHITE
- * @param bg background color, can assume values from BLACK to LIGHTGRAY
-*/
-void screenSetColor(screenColor fg, screenColor bg);
-
-#endif /* __SCREEN_H__ */
+#endif // SCREEN_H
